code/app/requirements.txt
streamlit>=1.26
langchain==0.1.11
langchain-community==0.0.27
langchain-core==0.1.30
langchain-text-splitters==0.0.1
chromadb==0.4.24
openai==1.42.0
python-dotenv==1.0.1
tiktoken==0.7.0